refactor(employees): reorganize file structure under employee-information (#363)
### TL;DR
Reorganized employee-related components into a dedicated `employee-information` directory for better code organization.

### What changed?
Moved employee-related components (delete-employee, employee-details, employee-form, employee-form-modal) into a new `employee-information` subdirectory and updated all corresponding import paths.

### How to test?
1. Navigate to the Personnel section
2. Verify that employee table loads correctly
3. Test employee CRUD operations:
   - Add a new employee
   - View employee details
   - Edit employee information
   - Delete an employee
4. Confirm that all employee-related functionality works as before

### Why make this change?
To improve code maintainability by grouping related employee components into a dedicated directory, making the codebase more organized and easier to navigate.
